Extend UIBarButtonItem Stepik items
