+++ Stepic/Legacy/Controllers/RegistrationSignIn/AuthSocial/SocialAuthViewController.swift
@@ -79,7 +79,7 @@ final class SocialAuthViewController: UIViewController {
     @IBOutlet weak var collectionView: UICollectionView!
     @IBOutlet weak var stepikLogoHeightConstraint: NSLayoutConstraint!
 
-    private lazy var closeBarButtonItem = UIBarButtonItem.closeBarButtonItem(
+    private lazy var closeBarButtonItem = UIBarButtonItem.stepikCloseBarButtonItem(
         target: self,
         action: #selector(self.onCloseClick(_:))
     )
